fix: Added second order to Swap transaction
Added second Order to Swap transaction orders array to make it apparent that a pair of orders should be included. It would be helpful to note that the ethSignature should be an array of ethSignatures [Swap, OrderA, OrderB] (rather than object for most other transactions) - the API response code is general and it's not easy to understand how it should be fixed.